renamed packages and classes as requested on review
