+++ sdks/java/io/amazon-web-services2/src/main/java/org/apache/beam/sdk/io/aws2/options/AwsOptions.java
@@ -18,9 +18,10 @@
 package org.apache.beam.sdk.io.aws2.options;
 
 import java.net.URI;
+import org.apache.beam.sdk.io.aws2.auth.StsAssumeRoleForFederatedCredentialsProvider;
+import org.apache.beam.sdk.io.aws2.auth.WebIdTokenProvider;
 import org.apache.beam.sdk.io.aws2.common.ClientBuilderFactory;
 import org.apache.beam.sdk.io.aws2.common.HttpClientConfiguration;
-import org.apache.beam.sdk.io.aws2.common.auth.providers.StsAssumeRoleWithDynamicWebIdentityCredentialsProvider;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.DefaultValueFactory;
 import org.apache.beam.sdk.options.Description;
@@ -147,17 +148,17 @@ public interface AwsOptions extends PipelineOptions {
    *   "durationSeconds": 3600
    * }}</pre>
    *
-   * <li>{@link StsAssumeRoleWithDynamicWebIdentityCredentialsProvider}
+   * <li>{@link StsAssumeRoleForFederatedCredentialsProvider}
    *
    *     <p>Similar to {@link StsAssumeRoleWithWebIdentityCredentialsProvider} but supporting
-   *     long-running jobs. The configuration expects the fqcn for an implementation of {@link
-   *     org.apache.beam.sdk.io.aws2.common.auth.providers.WebIdTokenProvider} which will be used to
-   *     retrieve a fresh token with each STS auth request.
+   *     long-running jobs. The configuration expects the FQCN for an implementation of {@link
+   *     WebIdTokenProvider} which will be used to retrieve a fresh token with each STS auth
+   *     request.
    *
    *     <pre>{@code --awsCredentialsProvider={
    *   "@type": "StsAssumeRoleWithDynamicWebIdentityCredentialsProvider",
    *   "roleArn": "role_arn_Value",
-   *   "roleSessionName": "session_name_value",
+   *   "audience": "audience_value",
    *   "webIdTokenProviderFQCN": "fully_qualified_class_name",
    *   "durationSeconds": 3600
    * }}</pre>
